+++ src/main/java/pl/plajer/murdermystery/events/Events.java
@@ -135,7 +135,7 @@ public class Events implements Listener {
               }
               if (victim.getLocation().distance(loc) < 1.0) {
                 if (!victim.equals(attacker)) {
-                  Utils.spawnCorpse(victim, arena);
+                  plugin.getCorpseHandler().spawnCorpse(victim, arena);
                   victim.damage(100.0);
                   victim.getWorld().playSound(victim.getLocation(), Sound.ENTITY_PLAYER_DEATH, 50, 1);
                   MessageUtils.sendTitle(victim, ChatManager.colorMessage("In-Game.Messages.Game-End-Messages.Titles.Died"), 5, 40, 5);