+++ src/main/java/pl/plajer/murdermystery/utils/Utils.java
@@ -65,25 +65,4 @@ public class Utils {
     }.runTaskTimer(plugin, 10, 10);
   }
 
-  public static void spawnCorpse(Player p, Arena arena) {
-    try {
-      Corpses.CorpseData corpse = CorpseAPI.spawnCorpse(p, p.getLocation());
-      Hologram hologram = HologramsAPI.createHologram(plugin, p.getLocation().clone().add(0, 1.5, 0));
-      hologram.appendTextLine(ChatManager.colorMessage("In-Game.Messages.Corpse-Last-Words").replace("%player%", p.getName()));
-      //todo priority note to wiki
-      if(p.hasPermission("murdermystery.lastwords.meme")) {
-        hologram.appendTextLine(ChatManager.colorMessage("In-Game.Messages.Last-Words.Meme"));
-      } else if(p.hasPermission("murdermystery.lastwords.rage")) {
-        hologram.appendTextLine(ChatManager.colorMessage("In-Game.Messages.Last-Words.Rage"));
-      } else if(p.hasPermission("murdermystery.lastwords.pro")) {
-        hologram.appendTextLine(ChatManager.colorMessage("In-Game.Messages.Last-Words.Pro"));
-      } else {
-        hologram.appendTextLine(ChatManager.colorMessage("In-Game.Messages.Last-Words.Default"));
-      }
-      arena.addCorpse(new ArenaCorpse(hologram, corpse));
-    } catch (Exception ex) {
-      new ReportedException(plugin, ex);
-    }
-  }
-
 }