+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * JavaDocs created
 * Fixed throwable sword angle
 * You cannot join game now infinitely through command
+* Added corpses override option in config
 
 ### 0.0.7 Beta (03.10.2018)
 * Added PAPI placeholders in chat (chat formatting must be enabled)